requirements.txt
pandas
numpy
sentence-transformers
langchain
qdrant-client
huggingface
streamlit
huggingface-hub
python-dotenv
openai
pyyaml
torch
langchain_community
tiktoken
pydub
gTTS
SpeechRecognition
PyAudio
beautifulsoup4
pypdf
docx
